Move TestGroupSpecificStartState to command_list_manager tests
