+++ pilz_trajectory_generation/test/integrationtest_sequence_action_capability.test
@@ -32,7 +32,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
     <arg name="fake_execution" value="true"/>
     <arg name="info" value="true"/>
     <arg name="debug" value="$(arg debug)"/>
-    <arg name="gripper" value="pg70"/>
     <arg name="pipeline" value="pilz_command_planner" />
   </include>
